requirements.txt
numpy
pandas 
torch 
torchvision 
torchaudio
faiss-cpu
sentence-transformers
requests
streamlit 
faiss-cpu
huggingface_hub
